Add samples from the posterior fit or posterior prediction of a model to a data frame

Given a data frame, adds samples from the posterior fit (aka the linear/link-level predictor) or the posterior predictions of the model to the data in a long format.

add_predicted_samples(newdata, model, var = "pred", ..., n = NULL)

add_fitted_samples(newdata, model, var = "estimate", ..., n = NULL)

predicted_samples(model, newdata, var = "pred", ..., n = NULL)

fitted_samples(model, newdata, var = "estimate", ..., n = NULL)

# S3 method for default
predicted_samples(model, newdata, ...)

# S3 method for default
fitted_samples(model, newdata, ...)

# S3 method for stanreg
predicted_samples(model, newdata, var = "pred", ...,
  n = NULL)

# S3 method for stanreg
fitted_samples(model, newdata, var = "estimate", ...,
  n = NULL)

# S3 method for brmsfit
predicted_samples(model, newdata, var = "pred", ...,
  n = NULL)

# S3 method for brmsfit
fitted_samples(model, newdata, var = "estimate", ...,
  n = NULL)

Arguments

newdata
Data frame to generate predictions from. If omitted, most model types will generate predictions from the data used to fit the model.

model
A supported Bayesian model fit / MCMC object. Currently supported models include mcmc, mcmc.list, runjags, stanfit, and anything with its own as.mcmc.list implementation. If you install the tidybayes.rethinking package (available at https://github.com/mjskay/tidybayes.rethinking), map and map2stan models from the rethinking package are also supported.

var
The name of the output column for the predictions (default `code"pred"`) or fits (default "estimate", for compatibility with tidy).

...
Additional arguments passed to the underlying prediction method for the type of model given.

n
The number of samples per prediction / fit to return.

Value

A data frame (actually, a tibble) with a .row column (a factor grouping rows from the input newdata), .chain column (the chain each sample came from, or NA if the model does not provide chain information), .iteration column (the iteration the sample came from), and .pred column (a sample from the posterior predictive distribution). For convenience, the resulting data frame comes grouped by the original input rows.

Details

add_fitted_samples adds samples from the posterior linear predictor (or the "link") to the data. It corresponds to posterior_linpred in rstanarm or fitted.brmsfit in brms. add_predicted_samples adds samples from the posterior prediction to the data. It corresponds to posterior_predict in rstanarm or predict.brmsfit in brms. add_fitted_samples and fitted_samples are alternate spellings of the same function with opposite order of the first two arguments to facilitate use in data processing pipelines that start either with a data frame or a model. Similarly, add_predicted_samples and predicted_samples are alternate spellings.

Given equal choice between the two, add_fitted_samples and add_predicted_samples are the preferred spellings.

See also

spread_samples

Examples

##TODO