Get samples from a model as a tibble

Extract samples from an MCMC fit into a wide-format data frame with a .chain and .iteration column, as well as all terms as columns. Generally speaking not as useful as spread_samples or gather_samples and is mainly used interally (see `Details`)

as_sample_tibble(model)

as_sample_data_frame(model)

# S3 method for default
as_sample_tibble(model)

# S3 method for mcmc.list
as_sample_tibble(model)

# S3 method for stanfit
as_sample_tibble(model)

# S3 method for stanreg
as_sample_tibble(model)

# S3 method for runjags
as_sample_tibble(model)

# S3 method for brmsfit
as_sample_tibble(model)

# S3 method for matrix
as_sample_tibble(model)

# S3 method for MCMCglmm
as_sample_tibble(model)

Arguments

model
A supported Bayesian model fit / MCMC object. See tidybayes-models for a list of supported models.

Value

A data frame (actually, a tibble) with a .chain column, .iteration column, and one column for every parameter in model.

Details

You will not typically call this directly; instead use spread_samples or gather_samples. However, to provide support for new models in those functions, you must provide an implementation of this function, or an implementaiton of as.mcmc.list (as_sample_tibble should work on any model with an implementation of as.mcmc.list) as_sample_data_frame is an alias.

See also

spread_samples or gather_samples, which use this function internally and provides a friendly interface for extracting tidy data frames from model fits.

Examples

library(magrittr)

data(line, package = "coda")

line %>%
  as_sample_tibble()

#> # A tibble: 400 x 5
#>    .chain .iteration alpha   beta  sigma
#>     <int>      <int> <dbl>  <dbl>  <dbl>
#>  1      1          1  7.17 -1.57  11.2  
#>  2      1          2  2.95  1.50   4.89 
#>  3      1          3  3.67  0.628  1.40 
#>  4      1          4  3.32  1.18   0.663
#>  5      1          5  3.71  0.490  1.36 
#>  6      1          6  3.58  0.207  1.04 
#>  7      1          7  2.70  0.883  1.29 
#>  8      1          8  2.96  1.09   0.459
#>  9      1          9  3.53  1.07   0.634
#> 10      1         10  2.09  1.48   0.913
#> # ... with 390 more rows