Bandwidth estimators

Bandwidth estimators for densities, used in the bandwidth argument to density functions (e.g. density_bounded(), density_unbounded()). Supports automatic partial function application.

bandwidth_nrd0(x, ...)

bandwidth_nrd(x, ...)

bandwidth_ucv(x, ...)

bandwidth_bcv(x, ...)

bandwidth_SJ(x, ...)

bandwidth_dpi(x, ...)

Arguments

x

A numeric vector giving a sample.

...

Arguments passed on to stats::bw.SJ

nb

number of bins to use.

lower,upper

range over which to minimize. The default is almost always satisfactory. hmax is calculated internally from a normal reference bandwidth.

method

either "ste" ("solve-the-equation") or "dpi" ("direct plug-in"). Can be abbreviated.

tol

for method "ste", the convergence tolerance for uniroot. The default leads to bandwidth estimates with only slightly more than one digit accuracy, which is sufficient for practical density estimation, but possibly not for theoretical simulation studies.

Value

A single number giving the bandwidth

Details

These are loose wrappers around the corresponding bw.-prefixed functions in stats. See, for example, bw.SJ().

bandwidth_dpi(), which is the default bandwidth estimator in ggdist, is the Sheather-Jones direct plug-in estimator, i.e. bw.SJ(..., method = "dpi").

See also