Estimate bounds of a distribution using Cooke's method

Estimate the bounds of the distribution a sample came from using Cooke's method. Use with the bounder argument to density_bounded().

Supports automatic partial function application with waived arguments.

Usage

bounder_cooke(x)

Arguments

x

<numeric> Sample to estimate the bounds of.

Value

A length-2 numeric vector giving an estimate of the minimum and maximum bounds of the distribution that x came from.

Details

Estimate the bounds of a distribution using the method from Cooke (1979); i.e. method 2.3 from Loh (1984). These bounds are:

$$\left[\begin{array}{l} 2X_{(1)} - \sum_{i = 1}^n \left[\left(1 - \frac{i - 1}{n}\right)^n - \left(1 - \frac{i}{n}\right)^n \right] X_{(i)}\\ 2X_{(n)} - \sum_{i = 1}^n \left[\left(1 - \frac{n - i}{n}\right)^n - \left(1 - \frac{n + 1 - i}{n} \right)^n\right] X_{(i)} \end{array}\right]$$

Where \(X_{(i)}\) is the \(i\)th order statistic of x (i.e. its \(i\)th-smallest value).

References

Cooke, P. (1979). Statistical inference for bounds of random variables. Biometrika 66(2), 367–374. doi:10.1093/biomet/66.2.367 .

Loh, W. Y. (1984). Estimating an endpoint of a distribution with resampling methods. The Annals of Statistics 12(4), 1543–1550. doi:10.1214/aos/1176346811

See also