Extract samples of lsmeans estimates from a Bayesian model fit into a tidy data format

Extract samples from the result of a call to lsmeans or ref.grid applied to a Bayesian model.

gather_lsmeans_samples(object)

Arguments

object
A ref.grid object such as returned by ref.grid or lsmeans.

Value

A data frame of tidy samples. The columns of the reference grid are returned as-is, with an additional column called estimate containing samples from the marginal estimates (the name estimate is used for compatibility with gather_samples and tidy). The resulting data frame is grouped by the columns from the reference grid to make use of summary functions like point_interval straightforward.

Details

lsmeans provides a convenient syntax for generating marginal estimates from a model, and can be applied to various Bayesian models, like stanreg-objects and MCMCglmm. Given a ref.grid object as returned by functions like ref.grid or lsmeans applied to a Bayesian model, gather_lsmeans_samples returns a tidy format data frame of samples from the marginal posterior distributions generated by lsmeans.

See also

lsmeans

Examples

##TODO