Gather terms from a tidy data frame of parameter samples into a single column

Given a data frame such as might be returned by as_sample_tibble or spread_samples, gather terms and their estimates from that data frame into a term and estimate column.

gather_terms(data, ignore_columns = "^\\..*")

Arguments

data
A data frame with parameter/term names spread across columns, such as one returned by as_sample_tibble or spread_samples.

ignore_columns
A regular expression that matches column names to ignore in the gather. The default ignores columns that start with ".".

Value

A data frame.

Details

This function gather every column except grouping columns and those matching the regular expression ignore_columns into key/value columns "term" and "estimate".

This function uses "term" and "estimate" instead of names like "parameter" and "value" in order to be consistent with the naming scheme of tidy.

Imagine a data frame data as returned by spread_samples(fit, a[i], b[i,v]), like this:

column ".chain": the chain number

column ".iteration": the interation number

column "i": value in 1:5

column "v": value in 1:10

column "a": value of "a[i]" for iteration number ".iteration" on chain number ".chain"

column "b": value of "b[i,v]" for iteration number ".iteration" on chain number ".chain"

gather_terms(data) on that data frame would return a grouped data frame (grouped by i and v), with:

column ".chain": the chain number

column ".iteration": the interation number

column "i": value in 1:5

column "v": value in 1:10

column "term": value in c("a", "b").

column "estimate": value of "a[i]" (when "term" is "a"; repeated for every value of "v") or "b[i,v]" (when "term" is "b") for iteration number ".iteration" on chain number ".chain"

In this example, this call:

gather_terms(data)

Is roughly equivalent to:

data %>%
  gather(term, estimate, -c(.chain, .iteration, i, v)) %>%
  group_by(term, add = TRUE)

See also

spread_samples, as_sample_tibble.

Examples

library(dplyr)
data(RankCorr)

RankCorr %>%
  spread_samples(b[i,v], tau[i]) %>%
  gather_terms() %>%
  mean_qi()

#> |===================================================   | 94% ~0 s remaining     |====================================================  | 97% ~0 s remaining     

#> # A tibble: 144 x 7
#> # Groups:   i, v, term [144]
#>        i     v  term    estimate    conf.low  conf.high .prob
#>    <int> <int> <chr>       <dbl>       <dbl>      <dbl> <dbl>
#>  1     1     1     b -1.07813380 -1.24631300 -0.9054563  0.95
#>  2     1     1   tau  6.03129167  5.07486825  7.1405325  0.95
#>  3     1     2     b -1.82516817 -2.07576900 -1.5438657  0.95
#>  4     1     2   tau  6.03129167  5.07486825  7.1405325  0.95
#>  5     1     3     b  0.17349259  0.03489905  0.3170186  0.95
#>  6     1     3   tau  6.03129167  5.07486825  7.1405325  0.95
#>  7     1     4     b -0.09884865 -0.34174593  0.1410597  0.95
#>  8     1     4   tau  6.03129167  5.07486825  7.1405325  0.95
#>  9     2     1     b -0.73395324 -1.03256875 -0.4346341  0.95
#> 10     2     1   tau  3.37725239  2.40876750  4.5755795  0.95
#> # ... with 134 more rows


# the first three lines below are roughly equivalent to ggmcmc::ggs(RankCorr)
RankCorr %>%
  as_sample_tibble() %>%
  gather_terms() %>%
  mean_qi()

#> |====================================                  | 68% ~1 s remaining     |======================================                | 71% ~1 s remaining     |======================================                | 72% ~1 s remaining     |=======================================               | 74% ~1 s remaining     |========================================              | 76% ~1 s remaining     |==========================================            | 78% ~1 s remaining     |==========================================            | 80% ~1 s remaining     |============================================          | 82% ~1 s remaining     |=============================================         | 84% ~0 s remaining     |==============================================        | 86% ~0 s remaining     |===============================================       | 88% ~0 s remaining     |================================================      | 89% ~0 s remaining     |=================================================     | 91% ~0 s remaining     |==================================================    | 93% ~0 s remaining     |===================================================   | 95% ~0 s remaining     |====================================================  | 97% ~0 s remaining     |===================================================== | 99% ~0 s remaining     

#> # A tibble: 235 x 5
#> # Groups:   term [235]
#>       term    estimate    conf.low  conf.high .prob
#>      <chr>       <dbl>       <dbl>      <dbl> <dbl>
#>  1  b[1,1] -1.07813380 -1.24631300 -0.9054563  0.95
#>  2  b[1,2] -1.82516817 -2.07576900 -1.5438657  0.95
#>  3  b[1,3]  0.17349259  0.03489905  0.3170186  0.95
#>  4  b[1,4] -0.09884865 -0.34174593  0.1410597  0.95
#>  5 b[10,1] -0.49728198 -0.67700202 -0.3091694  0.95
#>  6 b[10,2] -1.70815594 -1.99129300 -1.4242285  0.95
#>  7 b[10,3]  0.03443984 -0.11406630  0.1873531  0.95
#>  8 b[10,4]  0.32755347  0.07510573  0.5739541  0.95
#>  9 b[11,1] -1.19414856 -1.37085550 -1.0125650  0.95
#> 10 b[11,2] -2.38960803 -2.65112400 -2.1254285  0.95
#> # ... with 225 more rows