Eye plots (aka raindrop plots, aka violin plots) of densities with point estimates and intervals for ggplot2

Generates a combination geom_violin and stat_pointinterval (for geom_eye) or geom_violinh and stat_pointintervalh (for geom_eyeh) representing the density, point estimates, and uncertainty intervals for some samples. Useful for representing posterior estimates from Bayesian samplers; in that context this is variously called an eye plot, a raindrop plot, or a violin plot (though violin plot is also applied to plots of data, hence its use is not preferred here).

geom_eye(mapping = NULL, data = NULL, stat = "ydensity",
  position = "dodge", trim = TRUE, scale = "area", fill = "gray65", ...,
  point.interval = mean_qi, fun.data = point.interval, fun.args = list(),
  .prob = c(0.95, 0.66), color = NULL, size = NULL,
  fatten.interval = NULL, fatten.point = NULL)

geom_eyeh(mapping = NULL, data = NULL, stat = "xdensity",
  position = "dodgev", trim = TRUE, scale = "area", fill = "gray65",
  ..., point.interval = mean_qih, fun.data = point.interval,
  fun.args = list(), .prob = c(0.95, 0.66), color = NULL, size = NULL,
  fatten.interval = NULL, fatten.point = NULL)

Arguments

mapping
The aesthetic mapping, usually constructed with aes or aes_string. Only needs to be set at the layer level if you are overriding the plot defaults.

data
A layer specific dataset - only needed if you want to override the plot defaults.

stat
Passed to geom_violin / geom_violinh. The statistical transformation to use on the data for this layer.

position
Passed to geom_violin / geom_violinh. The position adjustment to use for overlapping points on this layer.

trim
Passed to geom_violin / geom_violinh. If TRUE (default), trim the tails of the violins to the range of the data. If FALSE, don't trim the tails.

scale
Passed to geom_violin / geom_violinh. If "area" (default), all violins have the same area (before trimming the tails). If "count", areas are scaled proportionally to the number of observations. If "width", all violins have the same maximum width.

fill
Passed to geom_violin / geom_violinh. Fill color of the violin.

...
Currently unused.

point.interval
Alias for fun.data

fun.data
A function that is given a vector and should return a data frame with variables y, ymin and ymax (x, xmin and xmax for stat_pointintervalh), and .prob. See the point_interval family of functions.

fun.args
Optional arguments passed to fun.data.

.prob
The .prob argument passed to fun.data.

color
Passed to stat_pointintervalh. Color of the point estimate and credible interval.

size
Passed to stat_pointintervalh. Line weight of the point estimate and credible interval.

fatten.interval
A multiplicative factor used to adjust the size of the interval lines (line size will be (size + 3) * fatten.interval. The default decreases the line size, because the default range of scale_size_continuous has an upper end of 6, which is quite large.

fatten.point
A multiplicate factor used to adjust the size of the point relative to the largest line.

Details

An eye plot is a compact visual summary of the distribution of some samples, used (under various names and with subtle variations) to visualize posterior distributions in Bayesian inference. This instantiation is a combination of a violin plot, point estimate, and credible interval.

The vertical form, geom_eye, is equivalent to geom_violin() + stat_pointinterval() with some reasonable defaults, including color choices and the use of mean with 95% and 6% quantile intervals.

The horizontal form, geom_eyeh(), is equivalent to geom_violinh() + stat_pointintervalh().

See also

See geom_halfeyeh for the non-mirrored density ("half eye") version. See geom_violin and stat_pointinterval for the geoms these functions are based on.

Examples

#TODO