Horizontal half-eye plots of densities with point estimates and intervals for ggplot2

Generates a combination geom_joy and geom_pointrangeh (using stat_summaryh) representing the density, point estimates, and credible interval. Useful for representing posterior estimates from Bayesian samplers; in that context the mirrored verison is variously called an eye plot, a raindrop plot, or a violin plot; hence "half-eye" for this plot.

geom_halfeyeh(mapping = NULL, data = NULL, stat = "joy",
  position = "identity", scale = 0.9, rel_min_height = 0,
  fill = "gray65", density.args = list(), ..., fun.data = mean_qih,
  fun.args = list(), color = NULL, size = NULL,
  interval.args = list(geom = "pointrangeh", position = "identity"))

Arguments

mapping
The aesthetic mapping, usually constructed with aes or aes_string. Only needs to be set at the layer level if you are overriding the plot defaults.

data
A layer specific dataset - only needed if you want to override the plot defaults.

stat
Passed to geom_joy. The statistical transformation to use on the data for this layer.

position
Passed to geom_joy. The position adjustment to use for overlapping points on this layer.

scale
Passed to geom_joy. A scaling factor to scale the height of the densities relative to the spacing between them. A value of 1 indicates that the maximum point of any density touches the baseline right above, assuming even spacing between baselines.

rel_min_height
Densities at heights below this cutoff will be removed. The cutoff is measured relative to the overall maximum, so rel_min_height=0.01 would remove everything that is 1 densities. Default is 0, so nothing is removed.

fill
Passed to geom_joy. Fill color of the density.

density.args
Other arguments passed to geom_joy.

...
Other arguments passed to geom_joy.

fun.data
The function used to construct point estimates and credible intervals from the data. Should be a function that takes a vector of data as input and returns a data frame with columns y, ymin, and ymax, representing the point estimate, lower bound of the credible interval, and upper bound of the credible interval. See point_interval for examples; some typical functions to consider here are mean_qi (mean with quantile interval), median_qi (median with quantile interval), and mode_hdi (mode with highest density interval). Alias for fun.data.

fun.args
Optional arguments passed to fun.data.

color
Passed to stat_summaryh. Color of the point estimate and credible interval.

size
Passed to stat_summaryh. Line weight of the point estimate and credible interval.

interval.args
Other arguments passed to stat_summaryh.

Details

A half-eye plot is a compact visual summary of the distribution of some samples, used (under various names and with subtle variations) to visualize posterior distributions in Bayesian inference. This instantiation is a combination of a density plot, point estimate, and credible interval. geom_halfeyeh() is equivalent to geom_joy() + stat_summaryh() with some reasonable defaults, including color choices and the use of mean with 95% quantile intervals.

See also

See geom_eye and geom_eyeh for the mirrored-density (full "eye") versions. See geom_joy and stat_summaryh for the geoms this function is based on.

Examples

##TODO