Prefix function generator for composing index columns

Generates a function for generating names of index columns for factors in compose_data by prefixing a character vector to the original column name.

n_prefix(prefix)

Arguments

prefix
Character vector to be prepended to column names by compose_data to create index columns. Typically something like "n" (that is the default used in the .n_name argument of compose_data).

Returns a function. The function returned takes a character vector, name and returns paste0(prefix, "_", name), unless name is empty, in which case it will return prefix. n_prefix("n") is the default method that compose_data uses to generate column names for factor indices. Under this method, given a data frame df with a factor column "foo" containing 5 levels, the results of compose_data(df) will include an element named "n" (the result of n_prefix("n")("")) equal to the number of rows in df and an element named "n_foo" (the result of n_prefix("n")("foo")) equal to the number of levels in df$foo.

See also

The .n_name argument of compose_data