Get the names of the parameters in a fitted Bayesian model

Get a character vector of the names of the parameters in a variety of fitted Bayesian models. All models supported by as_sample_tibble are supported.

parameters(model)

# S3 method for default
parameters(model)

Arguments

model
A supported Bayesian model fit / MCMC object. Currently supported models include mcmc, mcmc.list, runjags, stanfit, stanreg-objects, brm, and anything with its own as.mcmc.list implementation. If you install the tidybayes.rethinking package (available at https://github.com/mjskay/tidybayes.rethinking), map and map2stan models from the rethinking package are also supported.

Value

A character vector of parameter names in the fitted model.

Details

This function is often useful for inspecting a model interactively in order to construct calls to spread_samples or gather_samples in order to extract samples in a tidy format.

See also

spread_samples, gather_samples.

Examples

data(line, package = "coda")
parameters(line)

#> [1] "alpha" "beta"  "sigma"

data(RankCorr, package = "tidybayes")
parameters(RankCorr)

#>   [1] "b[1,1]"         "b[2,1]"         "b[3,1]"         "b[4,1]"        
#>   [5] "b[5,1]"         "b[6,1]"         "b[7,1]"         "b[8,1]"        
#>   [9] "b[9,1]"         "b[10,1]"        "b[11,1]"        "b[12,1]"       
#>  [13] "b[13,1]"        "b[14,1]"        "b[15,1]"        "b[16,1]"       
#>  [17] "b[17,1]"        "b[18,1]"        "b[1,2]"         "b[2,2]"        
#>  [21] "b[3,2]"         "b[4,2]"         "b[5,2]"         "b[6,2]"        
#>  [25] "b[7,2]"         "b[8,2]"         "b[9,2]"         "b[10,2]"       
#>  [29] "b[11,2]"        "b[12,2]"        "b[13,2]"        "b[14,2]"       
#>  [33] "b[15,2]"        "b[16,2]"        "b[17,2]"        "b[18,2]"       
#>  [37] "b[1,3]"         "b[2,3]"         "b[3,3]"         "b[4,3]"        
#>  [41] "b[5,3]"         "b[6,3]"         "b[7,3]"         "b[8,3]"        
#>  [45] "b[9,3]"         "b[10,3]"        "b[11,3]"        "b[12,3]"       
#>  [49] "b[13,3]"        "b[14,3]"        "b[15,3]"        "b[16,3]"       
#>  [53] "b[17,3]"        "b[18,3]"        "b[1,4]"         "b[2,4]"        
#>  [57] "b[3,4]"         "b[4,4]"         "b[5,4]"         "b[6,4]"        
#>  [61] "b[7,4]"         "b[8,4]"         "b[9,4]"         "b[10,4]"       
#>  [65] "b[11,4]"        "b[12,4]"        "b[13,4]"        "b[14,4]"       
#>  [69] "b[15,4]"        "b[16,4]"        "b[17,4]"        "b[18,4]"       
#>  [73] "pred_y[1]"      "pred_y[2]"      "pred_y[3]"      "pred_y[4]"     
#>  [77] "pred_y[5]"      "pred_y[6]"      "pred_y[7]"      "pred_y[8]"     
#>  [81] "pred_y[9]"      "pred_y[10]"     "pred_y[11]"     "pred_y[12]"    
#>  [85] "pred_y[13]"     "pred_y[14]"     "pred_y[15]"     "pred_y[16]"    
#>  [89] "pred_y[17]"     "pred_y[18]"     "pred_y[19]"     "pred_y[20]"    
#>  [93] "pred_y[21]"     "pred_y[22]"     "pred_y[23]"     "pred_y[24]"    
#>  [97] "pred_y[25]"     "pred_y[26]"     "pred_y[27]"     "pred_y[28]"    
#> [101] "pred_y[29]"     "pred_y[30]"     "pred_y[31]"     "pred_y[32]"    
#> [105] "pred_y[33]"     "pred_y[34]"     "pred_y[35]"     "pred_y[36]"    
#> [109] "pred_y[37]"     "pred_y[38]"     "pred_y[39]"     "pred_y[40]"    
#> [113] "pred_y[41]"     "pred_y[42]"     "pred_y[43]"     "pred_y[44]"    
#> [117] "pred_y[45]"     "pred_y[46]"     "pred_y[47]"     "pred_y[48]"    
#> [121] "pred_y[49]"     "pred_y[50]"     "pred_y[51]"     "pred_y[52]"    
#> [125] "pred_y[53]"     "pred_y[54]"     "pred_y[55]"     "pred_y[56]"    
#> [129] "pred_y[57]"     "pred_y[58]"     "pred_y[59]"     "pred_y[60]"    
#> [133] "pred_y[61]"     "pred_y[62]"     "pred_y[63]"     "pred_y[64]"    
#> [137] "pred_y[65]"     "pred_y[66]"     "pred_y[67]"     "pred_y[68]"    
#> [141] "pred_y[69]"     "pred_y[70]"     "pred_y[71]"     "pred_y[72]"    
#> [145] "pred_y[73]"     "pred_y[74]"     "pred_y[75]"     "pred_y[76]"    
#> [149] "pred_y[77]"     "pred_y[78]"     "pred_y[79]"     "pred_y[80]"    
#> [153] "pred_y[81]"     "pred_y[82]"     "pred_y[83]"     "pred_y[84]"    
#> [157] "pred_y[85]"     "pred_y[86]"     "pred_y[87]"     "pred_y[88]"    
#> [161] "pred_y[89]"     "pred_y[90]"     "pred_y[91]"     "pred_y[92]"    
#> [165] "pred_y[93]"     "pred_y[94]"     "pred_y[95]"     "pred_y[96]"    
#> [169] "pred_y[97]"     "pred_y[98]"     "pred_y[99]"     "pred_y[100]"   
#> [173] "pred_y[101]"    "pred_y[102]"    "pred_y[103]"    "pred_y[104]"   
#> [177] "pred_y[105]"    "pred_y[106]"    "pred_y[107]"    "pred_y[108]"   
#> [181] "tau[1]"         "tau[2]"         "tau[3]"         "tau[4]"        
#> [185] "tau[5]"         "tau[6]"         "tau[7]"         "tau[8]"        
#> [189] "tau[9]"         "tau[10]"        "tau[11]"        "tau[12]"       
#> [193] "tau[13]"        "tau[14]"        "tau[15]"        "tau[16]"       
#> [197] "tau[17]"        "tau[18]"        "typical_mu[1]"  "typical_mu[2]" 
#> [201] "typical_mu[3]"  "typical_mu[4]"  "typical_mu[5]"  "typical_mu[6]" 
#> [205] "typical_mu[7]"  "typical_mu[8]"  "typical_mu[9]"  "typical_mu[10]"
#> [209] "typical_mu[11]" "typical_mu[12]" "typical_mu[13]" "typical_mu[14]"
#> [213] "typical_mu[15]" "typical_mu[16]" "typical_mu[17]" "typical_mu[18]"
#> [217] "typical_r"      "u_tau[1]"       "u_tau[2]"       "u_tau[3]"      
#> [221] "u_tau[4]"       "u_tau[5]"       "u_tau[6]"       "u_tau[7]"      
#> [225] "u_tau[8]"       "u_tau[9]"       "u_tau[10]"      "u_tau[11]"     
#> [229] "u_tau[12]"      "u_tau[13]"      "u_tau[14]"      "u_tau[15]"     
#> [233] "u_tau[16]"      "u_tau[17]"      "u_tau[18]"