Point and interval estimates for tidy sample data

Translates samples in a (possibly grouped) data frame into a point and interval estimate (or set of point and interval estimates, if there are multiple groups in a grouped data frame).

point_interval(.data, ..., .prob = 0.95, .point = mean, .interval = qi,
  .broom = TRUE)

# S3 method for default
point_interval(.data, ..., .prob = 0.95, .point = mean,
  .interval = qi, .broom = TRUE)

# S3 method for numeric
point_interval(.data, ..., .prob = 0.95, .point = mean,
  .interval = qi, .broom = FALSE)

point_intervalh(...)

qi(x, .prob = 0.95)

hdi(x, .prob = 0.95)

mean_qi(.data, ..., .prob = 0.95)

mean_qih(...)

median_qi(.data, ..., .prob = 0.95)

median_qih(...)

mode_qi(.data, ..., .prob = 0.95)

mode_qih(...)

mean_hdi(.data, ..., .prob = 0.95)

mean_hdih(...)

median_hdi(.data, ..., .prob = 0.95)

median_hdih(...)

mode_hdi(.data, ..., .prob = 0.95)

mode_hdih(...)

Arguments

.data
Data frame (or grouped data frame as returned by group_by) that contains samples to summarize.

...
Bare column names or expressions that, when evaluated in the context of .data, represent samples to summarise. If this is empty, then by default all columns that are not group columns or start with "." (e.g. ".chain" or ".iteration") will be summarised.

.prob
vector of probabilities to use for generating intervals. If multiple probabilities are provided, multiple rows per group are generated, each with a different probabilty interval (and value of the corresponding .prob column).

.point
Point estimate function, which takes a vector and returns a single value, e.g. mean, median, or Mode.

.interval
Interval estimate function, which takes a vector and a probability (.prob) and returns a two-element vector representing the lower and upper bound of an interval; e.g. qi, hdi

.broom
When TRUE and only a single estimate is to be output, use the name conf.low for the lower end of the interval and conf.high for the upper end for consistency with tidy in the broom package. If .data is a vector and this is TRUE, this will also set the column name of the point estimate to estimate.

x
vector to summarise (for interval functions: qi and hdi)

Details

If .data is a data frame, then ... is a list of bare names of columns (or expressions derived from columns) of .data, on which the point and interval estimates are derived. Column expressions are processed using the tidy evaluation framework (see eval_tidy).

For a column named x, the resulting data frame will have columns x (the point estimate), x.low (the lower end of the interval), x.high (the upper end of the interval), and .prob.

If .data includes groups (see e.g. group_by), the points and intervals are calculated within the groups.

If .data is a vector, ... is ignored and the result is a data frame with one row per value of .prob and three columns: y (the point estimate), ymin (the lower end of the interval), ymax (the upper end of the interval), and .prob, the probability corresponding to the interval. This behavior allows point_interval and its derived functions (like mean_qi, median_qi, etc) to be easily used to plot intervals in ggplot using methods like geom_eye, geom_eyeh, or stat_summary.

The functions ending in h (e.g., point_intervalh, mean_qih) behave identically to the function without the h, except that when passed a vector, they return a data frame with x/xmin/xmax instead of y/ymin/ymax. This allows them to be used as values of the fun.data = argument of stat_summaryh. mean_qi, mode_qi, etc are short forms for point_interval(..., .point = mean, .interval = qi), etc. qi yields the quantile interval (also known as the percentile interval or equi-tailed interval). hdi yields the highest-density interval (also known as the highest posterior density interval).

Examples

##TODO