Decorate a model fit or samples with data types recovered from the input data

Decorate the samples returned from a Bayesian sampler with types for variable and index data types. Meant to be used before calling spread_samples or gather_samples so that the values returned by those functions are translated back into useful data types.

recover_types(model, ...)

Arguments

model
A supported Bayesian model fit / MCMC object. Currently supported models include mcmc, mcmc.list, runjags, stanfit, stanreg-objects, brm, and anything with its own as.mcmc.list implementation. If you install the tidybayes.rethinking package (available at https://github.com/mjskay/tidybayes.rethinking), map and map2stan models from the rethinking package are also supported.

...
Lists (or data frames) providing data prototypes used to convert columns returned by spread_samples and gather_samples back into useful data types. See `Details`.

Value

A decorated version of model.

Details

Each argument in ... specifies a list or data.frame. The model is decorated with a list of constructors that can convert a numeric column into the data types in the lists in ....

Then, when spread_samples or gather_samples is called on the decorated model, each list entry with the same name as the variable or an index in varible_spec is a used as a prototype for that variable or index --- i.e., its type is taken to be the expected type of that variable or index. Those types are used to translate numeric values of variables back into useful values (for example, levels of a factor).

The most common use of recover_types is to automatically translate indices that correspond to levels of a factor in the original data back into levels of that factor. The simplest way to do this is to pass in the data frame from which the original data came.

Supported types of prototypes are factor, ordered, and logical. For example:

if prototypes$v is a factor, the v column in the returned samples is translated into a factor using factor(v, labels=levels(prototypes$v), ordered=is.ordered(prototypes$v)).

if prototypes$v is a logical, the v column is translated into a logical using as.logical(v).

Additional data types can be supported by providing a custom implementation of the generic function as_constructor.

See also

spread_samples, gather_samples, compose_data.

Examples

##TODO