Empty sub-guide for thickness scales

This is a blank sub-guide that omits annotations for the thickness and dot-count sub-scales in ggdist. It can be used with the subguide parameter of geom_slabinterval() and geom_dotsinterval().

Supports automatic partial function application with waived arguments.

Usage

subguide_none(values, ...)

Arguments

values

<numeric> Values used to construct the scale used for this guide. Typically provided automatically by geom_slabinterval().

...

ignored.

See also