Simple, light ggplot2 theme for ggdist and tidybayes

A simple, relatively minimalist ggplot2 theme, and some helper functions to go with it.

Usage

theme_ggdist(
  base_size = 11,
  base_family = "",
  base_line_size = base_size/22,
  base_rect_size = base_size/22
)

theme_tidybayes(
  base_size = 11,
  base_family = "",
  base_line_size = base_size/22,
  base_rect_size = base_size/22
)

facet_title_horizontal()

axis_titles_bottom_left()

facet_title_left_horizontal()

facet_title_right_horizontal()

Arguments

base_size

base font size, given in pts.

base_family

base font family

base_line_size

base size for line elements

base_rect_size

base size for rect elements

Value

A named list in the format of ggplot2::theme()

Details

This is a relatively minimalist ggplot2 theme, intended to be used for making publication-ready plots. It is currently based on ggplot2::theme_light().

A word of warning: this theme may (and very likely will) change in the future as I tweak it to my taste.

theme_ggdist() and theme_tidybayes() are aliases.

See also

Author

Matthew Kay

Examples

library(ggplot2)

theme_set(theme_ggdist())