Models supported by tidybayes

Tidybayes supports two classes of models and sample formats: Models that provide prediction functions, and those that do not.

All Supported Models

All supported models support the base tidybayes sample extraction functions, such as as_sample_tibble, spread_samples, and gather_samples. These models include:

rstan models

runjags models

brm models

rstanarm models

MCMCglmm models

mcmc and mcmc.list objects, which are output by several of model types.

Any object with an implementation of as.mcmc.list. For a list of those available in your environment, run methods(as.mcmc.list)

Models Supporting Prediction

In addition, the following models support fit and prediction extraction functions, such as add_fitted_samples and add_predicted_samples:

brm models

rstanarm models

If you install the tidybayes.rethinking package, models from the rethinking package are also supported. Note that in tidybayes.rethinking, tidy_link takes the place of add_fitted_samples and tidy_sim takes the place of add_predicted_samples.

Extending tidybayes

To include basic support for new models, one need only implement the as_sample_tibble generic function for that model.

To include support for estimation and prediction, one must implement the fitted_samples and predicted_samples generic functions.