Turn tidy data frames of parameters from a Bayesian model back into untidy data

Inverse operations of spread_samples and gather_samples, giving results that look like as_sample_tibble.

unspread_samples(data, ..., indices = c(".chain", ".iteration"),
  drop_indices = FALSE)

ungather_samples(data, ..., term = "term", estimate = "estimate",
  indices = c(".chain", ".iteration"), drop_indices = FALSE)

Arguments

data
A tidy data frame of samples, such as one output by spread_samples or gather_samples.

...
Expressions in the form of variable_name[index_1, index_2, ...]. See spread_samples.

indices
Character vector of column names in data that should be treated as indices of chain/iteration. The default is c(".chain",".iteration"), which are the same names used for chain/iteration indices returned by spread_samples or gather_samples.

drop_indices
Drop the columns specified by indices from the resulting data frame. Default FALSE.

term
The name of the column in data that contains the names of parameters from the model.

estimate
The name of the column in data that contains the samples of the parameters.

Value

A data frame.

Details

These functions take symbolic specifications of parameter names and indices in the same format as spread_samples and gather_samples and invert the tidy data frame back into a data frame whose column names are parameters with indices in them.

See also

spread_samples, gather_samples, as_sample_tibble.

Examples

##TODO