Decorate a model fit or samples with data types recovered from the input data

Decorate the samples returned from a Bayesian sampler with types for variable and index data types. Meant to be used before calling spread_samples or gather_samples so that the values returned by those functions are translated back into useful data types.

recover_types(model, ...)

Arguments

model
A supported Bayesian model fit / MCMC object. Currently supported models include mcmc, mcmc.list, runjags, stanfit, stanreg-objects, brm, and anything with its own as.mcmc.list implementation. If you install the tidybayes.rethinking package (available at https://github.com/mjskay/tidybayes.rethinking), map and map2stan models from the rethinking package are also supported.

...
Lists (or data frames) providing data prototypes used to convert columns returned by spread_samples and gather_samples back into useful data types. See `Details`.

Value

A decorated version of model.

Details

Each argument in ... specifies a list or data.frame. The model is decorated with a list of constructors that can convert a numeric column into the data types in the lists in ....

Then, when spread_samples or gather_samples is called on the decorated model, each list entry with the same name as the variable or an index in varible_spec is a used as a prototype for that variable or index --- i.e., its type is taken to be the expected type of that variable or index. Those types are used to translate numeric values of variables back into useful values (for example, levels of a factor).

The most common use of recover_types is to automatically translate indices that correspond to levels of a factor in the original data back into levels of that factor. The simplest way to do this is to pass in the data frame from which the original data came.

Supported types of prototypes are factor, ordered, and logical. For example:

if prototypes$v is a factor, the v column in the returned samples is translated into a factor using factor(v, labels=levels(prototypes$v), ordered=is.ordered(prototypes$v)).

if prototypes$v is a logical, the v column is translated into a logical using as.logical(v).

Additional data types can be supported by providing a custom implementation of the generic function as_constructor.

See also

spread_samples, gather_samples, compose_data.

Examples

library(dplyr)
library(magrittr)
library(rstan)

# Here's an example dataset with a categorical predictor (`condition`) with several levels:
set.seed(5)
n = 10
n_condition = 5
ABC =
  data_frame(
    condition = rep(c("A","B","C","D","E"), n),
    response = rnorm(n * 5, c(0,1,2,1,-1), 0.5)
  )

# We'll fit the following model to it:
stan_code = "
  data {
    int<lower=1> n;
    int<lower=1> n_condition;
    int<lower=1, upper=n_condition> condition[n];
    real response[n];
  }
  parameters {
    real overall_mean;
    vector[n_condition] condition_zoffset;
    real<lower=0> response_sd;
    real<lower=0> condition_mean_sd;
  }
  transformed parameters {
    vector[n_condition] condition_mean;
    condition_mean = overall_mean + condition_zoffset * condition_mean_sd;
  }
  model {
    response_sd ~ cauchy(0, 1);       // => half-cauchy(0, 1)
    condition_mean_sd ~ cauchy(0, 1); // => half-cauchy(0, 1)
    overall_mean ~ normal(0, 5);
    condition_zoffset ~ normal(0, 1); //=> condition_mean ~ normal(overall_mean, condition_mean_sd)
    for (i in 1:n) {
      response[i] ~ normal(condition_mean[condition[i]], response_sd);
    }
  }
  "

m = stan(model_code = stan_code, data = compose_data(ABC), control = list(adapt_delta=0.99),
  # 1 chain / few iterations just so example runs quickly
  # do not use in practice
  chains = 1, iter = 500)

#> In file included from C:/R/R-3.4.2/library/BH/include/boost/config.hpp:39:0,
#>                  from C:/R/R-3.4.2/library/BH/include/boost/math/tools/config.hpp:13,
#>                  from C:/R/R-3.4.2/library/StanHeaders/include/stan/math/rev/core/var.hpp:7,
#>                  from C:/R/R-3.4.2/library/StanHeaders/include/stan/math/rev/core/gevv_vvv_vari.hpp:5,
#>                  from C:/R/R-3.4.2/library/StanHeaders/include/stan/math/rev/core.hpp:12,
#>                  from C:/R/R-3.4.2/library/StanHeaders/include/stan/math/rev/mat.hpp:4,
#>                  from C:/R/R-3.4.2/library/StanHeaders/include/stan/math.hpp:4,
#>                  from C:/R/R-3.4.2/library/StanHeaders/include/src/stan/model/model_header.hpp:4,
#>                  from filef57466f312b6.cpp:8:
#> C:/R/R-3.4.2/library/BH/include/boost/config/compiler/gcc.hpp:186:0: warning: "BOOST_NO_CXX11_RVALUE_REFERENCES" redefined
#>  #  define BOOST_NO_CXX11_RVALUE_REFERENCES
#>  ^
#> <command-line>:0:0: note: this is the location of the previous definition
#> cc1plus.exe: warning: unrecognized command line option "-Wno-ignored-attributes"
#> 
#> SAMPLING FOR MODEL 'dd2cc433974aea50803799100a5d323d' NOW (CHAIN 1).
#> 
#> Gradient evaluation took 0 seconds
#> 1000 transitions using 10 leapfrog steps per transition would take 0 seconds.
#> Adjust your expectations accordingly!
#> 
#> 
#> Iteration:   1 / 500 [  0%]  (Warmup)
#> Iteration:  50 / 500 [ 10%]  (Warmup)
#> Iteration: 100 / 500 [ 20%]  (Warmup)
#> Iteration: 150 / 500 [ 30%]  (Warmup)
#> Iteration: 200 / 500 [ 40%]  (Warmup)
#> Iteration: 250 / 500 [ 50%]  (Warmup)
#> Iteration: 251 / 500 [ 50%]  (Sampling)
#> Iteration: 300 / 500 [ 60%]  (Sampling)
#> Iteration: 350 / 500 [ 70%]  (Sampling)
#> Iteration: 400 / 500 [ 80%]  (Sampling)
#> Iteration: 450 / 500 [ 90%]  (Sampling)
#> Iteration: 500 / 500 [100%]  (Sampling)
#> 
#>  Elapsed Time: 0.232 seconds (Warm-up)
#>                0.136 seconds (Sampling)
#>                0.368 seconds (Total)
#> 

# without using recover_types(), the `condition` column returned by spread_samples()
# will be an integer:
m %>%
  spread_samples(condition_mean[condition]) %>%
  mean_qi()

#> # A tibble: 5 x 5
#> # Groups: condition [5]
#>   condition condition_mean conf.low conf.high .prob
#>       <int>          <dbl>    <dbl>     <dbl> <dbl>
#> 1         1          0.199   -0.124     0.628 0.950
#> 2         2          0.996    0.653     1.33  0.950
#> 3         3          1.83     1.45      2.21  0.950
#> 4         4          1.03     0.658     1.39  0.950
#> 5         5         -0.883   -1.22     -0.534 0.950

# If we apply recover_types() first, subsequent calls to other tidybayes functions will
# automatically back-convert factors so that they are labeled with their original levels
# (assuming the same name is used)
m %<>% recover_types(ABC)

# now the `condition` column with be a factor with levels "A", "B", "C", ...
m %>%
  spread_samples(condition_mean[condition]) %>%
  mean_qi()

#> # A tibble: 5 x 5
#> # Groups: condition [5]
#>   condition condition_mean conf.low conf.high .prob
#>   <fctr>             <dbl>    <dbl>     <dbl> <dbl>
#> 1 A                  0.199   -0.124     0.628 0.950
#> 2 B                  0.996    0.653     1.33  0.950
#> 3 C                  1.83     1.45      2.21  0.950
#> 4 D                  1.03     0.658     1.39  0.950
#> 5 E                 -0.883   -1.22     -0.534 0.950