Simple, light ggplot2 theme for ggdist

A simple, relatively minimalist ggplot2 theme, and some helper functions to go with it.

theme_ggdist()

facet_title_left_horizontal()

facet_title_right_horizontal()

axis_titles_bottom_left()

Value

A ggplot2 theme

This is a relatively minimalist ggplot2 theme, intended to be used for making publication-ready plots. It is currently based on ggplot2::theme_light().

A word of warning: this theme may (and very likely will) change in the future as I tweak it to my taste.

See also

Examples

library(ggplot2)

theme_set(theme_ggdist())